add app core
