+++ src/bot/models_pydantic/users_questions.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from pydantic import BaseModel, EmailStr, Extra, Field, validator
 
-PHONE = r"^[\+]?[(]?[0-9]{3}[)]?[-\s\.]?[0-9]{3}[-\s\.]?[0-9]{4,6}$"
+from bot.validators import PHONE
 
 
 class UserContacts(BaseModel):
